Finished the app
Finished building the app with the additional calls out to a third-party forecasting service. Took me a while to get the hang of Vue because the documentation was a little wonky, but once I got it I dug it. Added handling on the server to fetch data on behalf of the client.